New small latin font; small font bug fix to md380-gfx
md380-gfx bug fix - height stored in table was 2x what it should have been for small fonts (12 instead of 6). I apparently missed this in my binary testing, but the software in the unit seemed to work.

small latin characters - performed serifectomy on letters, using the Apple II font as a general reference. I am unable to test with an actual unit; feedback on how the letters look is requested!