Better wording for ids that are not in users.csv
Sometimes (like TG 9998 Parrot for example) calls come in that are not
in the dmr-marc database. Other times new users arrive that aren't in
our loaded database yet. We shouldn't imply that the users.csv is not
loaded correctly. We should just state that the user isn't found in the
memory.