Fix for issue #712, cleanup of keyb.
The reported defect ('Menu Button Stops Working With Experimental "ON"') was isolated to a change to keyb.c made in commit fdc839e48c71ce9695328ec4c8f3dc0c2b366b20 ("allow calling up the menu induring rx").

The problem is two-fold:
1. The code does not actually work for the stated purpose (as the comment in the commit indicates).
2. The conditional guarding the code allows it to execute outside of the condition it was crafted for (aka: it triggers regardless of the idle/rx state of the radio).

Fixing problem 1 will require more thought.  This commit fixes problem 2 by limiting the executio of this code to when the radio is in the rx state.

While investigating the behavior in this code path, I performed some cleanup in the form of creating symbolic constants for the keycodes.  This reduces the number of magic numbers in the file, but not all of them.