Fix for issue #645
This is the minimal fix for this issue.  Before this fix, there were only two places where a corrupted file could show up:
1. From a curl download that fails without an error code
2. md380-fw generating a corrupted file

The other recipes in the firmware directory all check the size of the target and delete it and fail if there is a mismatch.  This minimal fix fills gap number 1 from above with a simple check for greater than zero bytes.