Added warning checks to ci, fixed another warning.
This commit is an attempt to stop the inflow of new warnings being added to the applet, while avoiding having to resolve all the #warning directives already in the codebase.

With this change, regular builds are unaffected; they will proceed as they have in the past.  "make ci", however, will promote warnings to errors, but #warnings will be supressed altogether.  The end result of this will be to make the ci fail when someone introduces a new warning to the applet.

This commit may warrant the closing of issue #704.