Test for KD4Z VM compilation. Modified SysTick_Handler
For some reason, the code generated on KD4Z's VM still is not the same
as when built elsewhere (including a 64-bit Debian system, and Windows
64-bit + MinGW).
This commit tries to fix it by letting SysTick_Handler "wait" until the
original firmware has called all necessary function.
Only after GPIO was initialized, the config was loaded, the
draw-statusline-hook was called, etc etc, SysTick_Handler now tries to
drive the backlight, and poll keys for the 'red button menu' aka 'app
menu'.

In older versions, SysTick_Handler basically did not do anything (but
call the original SysTick_Handler) until a certain number of
SysTick-interrupts was counted.

The calling sequence is now listed in what used to be Netmon3 (can be
inspected in the application menu):
- after 334 SysTicks, GPIO was initialized (by Tytera)
- after 363 SysTicks, the config-loading hook said "booting"
- after 4386 SysTicks, the status-line-drawing hook was called, etc. Why
these times depend on the host system on which the cross-compilation
toolchain runs (and why the toolchain generates a different binary code)
remains yet to be discovered.
The differences can be seen in the map file ( applet/applet.map) :
Addresses of our own functions and variables depend were different. Got
to find out one day why this happens !