Read mfg string before using USB extensions
This is what triggers hooking the DFU to enable additional commands.
It seems that on Linux at least, this string is read by default, but
at least on FreeBSD, it is not.

If the string is not accessed, the SPI ID is read as garbage and
the additional commands don't work.  Simply accessing it fixes this.