dmr-marc -> radioid.net
The dmr-marc databases are no longer available on
dmr-marc.net and dmr-marc.net points to radioid.net as
the new home.  Also added in some logging to get_dmrmarc_json.py.
It looks like there are bigger issues with the script, but they
are no longer being masked with a catch-all except/pass.